use strict fields, safer timeouts
i was passing the same number for the fetch response timeout and
the fetch request max_wait_time. this is wrong: max_wait_time is the
amount of time we want kafka to wait before sending its response when
there are no records yet available, so it should always be significantly
less than the timeout.